src/requirements.txt
bmp280
smbus
wiringpi
pyserial
pimoroni-bme280